INDY-1589: finish unit tests test_sign_validation_for_key_proof_exist.py
