+++ plenum/bls/bls_key_register_pool_manager.py
@@ -1,6 +1,5 @@
 from crypto.bls.bls_key_register import BlsKeyRegister
 from plenum.common.constants import BLS_KEY, ALIAS
-from plenum import config
 from plenum.common.constants import BLS_KEY, BLS_KEY_PROOF
 from plenum.server.pool_manager import TxnPoolManager
 
@@ -24,7 +23,7 @@ class BlsKeyRegisterPoolManager(BlsKeyRegister):
             self._current_pool_state_root_hash = pool_state_root_hash
             self._load_keys_for_root(pool_state_root_hash)
         return self._current_bls_keys.get(node_name) \
-            if not config.VALIDATE_SIGN_WITHOUT_BLS_KEY_PROOF or \
+            if self._pool_manager.config.VALIDATE_SIGN_WITHOUT_BLS_KEY_PROOF or \
                self._current_bls_keys_proof.get(node_name, None) not in [False, None] \
             else None
